factor out locale modules
