+++ inlang/packages/paraglide-js/src/compiler/output-structure/message-modules.test.ts
@@ -1,0 +1,85 @@
+import { test, expect } from "vitest";
+import { generateMessageModules } from "./message-modules.js";
+import type { Bundle, Message, ProjectSettings } from "@inlang/sdk";
+import type { CompiledBundleWithMessages } from "../compileBundle.js";
+
+test("should emit per locale message files", () => {
+	const resources: CompiledBundleWithMessages[] = [
+		{
+			bundle: {
+				code: 'console.log("bundle code");',
+				node: {
+					id: "happy_elephant",
+				} as unknown as Bundle,
+			},
+			messages: {
+				en: {
+					code: 'console.log("message in English");',
+					node: {} as unknown as Message,
+				},
+				de: {
+					code: 'console.log("message in German");',
+					node: {} as unknown as Message,
+				},
+			},
+		},
+	];
+
+	const settings: Pick<ProjectSettings, "locales" | "baseLocale"> = {
+		locales: ["en", "de"],
+		baseLocale: "en",
+	};
+
+	const fallbackMap: Record<string, string | undefined> = {
+		en: "de",
+		de: "en",
+	};
+
+	const output = generateMessageModules(
+		resources,
+		settings,
+		fallbackMap
+	);
+
+	expect(output).toHaveProperty("messages/en.js");
+	expect(output).toHaveProperty("messages/de.js");
+	expect(output).toHaveProperty("messages/en/happy_elephant.js");
+	expect(output).toHaveProperty("messages/de/happy_elephant.js");
+});
+
+test("the files should include files for each locale, even if there are no messages", () => {
+	const resources: CompiledBundleWithMessages[] = [
+		{
+			bundle: {
+				code: 'console.log("bundle code");',
+				node: {
+					id: "happy_elephant",
+				} as unknown as Bundle,
+			},
+			messages: {},
+		},
+	];
+
+	const settings: Pick<ProjectSettings, "locales" | "baseLocale"> = {
+		locales: ["en", "de", "fr"],
+		baseLocale: "en",
+	};
+
+	const fallbackMap: Record<string, string | undefined> = {
+		en: "de",
+		de: "en",
+	};
+
+	const output = generateMessageModules(
+		resources,
+		settings,
+		fallbackMap
+	);
+
+	expect(output).toHaveProperty("messages/en.js");
+	expect(output).toHaveProperty("messages/de.js");
+	expect(output).toHaveProperty("messages/fr.js");
+	expect(output).toHaveProperty("messages/en/happy_elephant.js");
+	expect(output).toHaveProperty("messages/de/happy_elephant.js");
+	expect(output).toHaveProperty("messages/fr/happy_elephant.js");
+});
